requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.1
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jedi==0.18.1
msgpack==1.0.3
parso==0.8.3
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pyheif==0.7.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
Wand==0.6.7
whitenoise==5.3.0
zope.interface==5.4.0
